feat : 더이상 행동을 할 수 없는 끝난 상태 추가
