+++ src/main/java/domain/state/Finished.java
@@ -1,0 +1,30 @@
+package domain.state;
+
+import domain.card.Card;
+import domain.user.Hand;
+
+import java.util.List;
+
+public abstract class Finished implements State {
+
+    private final Hand hand;
+
+    public Finished(Hand hand) {
+        this.hand = hand;
+    }
+
+    @Override
+    public State draw(Card card) {
+        throw new IllegalStateException("카드를 뽑을 수 없는 상태");
+    }
+
+    @Override
+    public State stay() {
+        throw new IllegalStateException("게임이 끝났습니다");
+    }
+
+    @Override
+    public List<Card> cards() {
+        return hand.getCards();
+    }
+}
