+++ src/main/java/domain/state/State.java
@@ -8,5 +8,7 @@ public interface State {
 
     State draw(Card card);
 
+    State stay();
+
     List<Card> cards();
 }
